pyproject.toml
[project]
name = "nlomop"
version = "0.1.0"
description = "Natural Language OMOP — conversational EHR query agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "openai>=1.60",
    "psycopg[binary,pool]>=3.2",
    "scipy>=1.14",
    "lifelines>=0.30",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
